refactor(rbac): merge instance_*.go + space_*.go into rbac.go and can.go
After Phase 5 of #330 the instance- and space-tier RBAC was already a
single SERVER_RBAC store with two parallel API names sharing the same
engine. The file split (instance_rbac.go vs space_rbac.go,
instance_can.go vs space_can.go) was the last visual remnant of the
dual-tier era.

Merge:
- instance_rbac.go (670 lines) + space_rbac.go (995 lines) → rbac.go
- instance_can.go (84 lines) + space_can.go (163 lines) → can.go

The merged files are organized by concern:
- rbac.go: SystemActorID const, hierarchy errors, RoleWithPermissions,
  init (initInstanceRBAC + CreateDefaultRoles), permission checking,
  server-tier assignment / permission ops / role CRUD, space-tier
  wrappers (auth-gated), room-level wrappers, and helper queries.
- can.go: server-tier Can* helpers, then space-tier admin/member Can*
  helpers, then room-scoped Can* helpers.

The matching test files are renamed to keep their git history
(rbac_test.go, rbac_space_test.go, can_test.go, can_space_test.go).
No public API changes; no behaviour changes.
