+++ cli/internal/core/can_test.go
